requirements.txt
python-sat>=0.1.8.dev13
Pillow>=10.0.0
